scraping/requirements.txt
# Web Scraping Data Pipeline Requirements

# Web scraping
playwright>=1.40.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# AI Enhancement
google-generativeai>=0.3.0

# Data validation and models
pydantic>=2.5.0

# Configuration
pyyaml>=6.0.1

# Async HTTP
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
